net: rocker: add common elements of pipeline model for rocker switch
This adds rocker support for the net_flow_get_* operations. With this
we can interrogate rocker for some of the basic objects it supports.

Here we see that for static configurations enabling the get operations
is simply a matter of defining a pipeline model and returning the
structures for the core infrastructure to encapsulate into netlink
messages.

Next we need to add inter-table metadata. This is not exposed in this
patch but is a common switch construct. A typical usage is to have a
table that can match on a set of fields and an action to set some
metadata. Then we can have a table later in the pipeline to match
on the metadata and implement a chain of actions.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>