net: flow_table: create a set of common headers and actions
This adds common headers, metadata and actions that drivers
can use. Drivers then need to export which headers, metadata,
and actions they will support and how they are linked in the
hardware.

This standardize the action space and metadata so that we
require drivers to export only valid actions and metadata.

The header space can be extended by the devices to include
device specific headers that are sometimes used by switch
devices. However by exposing the headers the complete layout
of the header must be provided as well as where it fits in
the packet.

Also the inter-table metadata fields may be non-obvious. The
inter-table metadata is an out of band data that gets passed
with a packet through the devices pipeline. It can typically
be matched on or set with a specific action. A common pattern
for switch devices is to use one table to set the inter-table
metadata and a subsequent table in the pipeline to match on
it and run some set of actions.

We allow switch drivers to specify instances of this inter-table
metadata as a special type of field. The inter-table meta-data
instances are scoped to the pipeline they are not exposed outside
the pipeline. Further there are no side effects from setting
the inter-table metadata and exiting the pipeline. What we
mean by this is if a table sets the inter-table value and a
packet exits the pipeline the device must not use the meta-data
for any forwarding, qos, or other decision.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>